+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/kotlin/com/microsoft/azure/toolkit/intellij/legacy/function/runner/localRun/profileStates/FunctionRunProfileState.kt
@@ -2,7 +2,7 @@
  * Copyright 2018-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the MIT license.
  */
 
-package com.microsoft.azure.toolkit.intellij.legacy.function.runner.localRun
+package com.microsoft.azure.toolkit.intellij.legacy.function.runner.localRun.profileStates
 
 import com.intellij.execution.DefaultExecutionResult
 import com.intellij.execution.ExecutionResult