+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/kotlin/com/microsoft/azure/toolkit/intellij/legacy/function/templates/InstallFunctionToolComponent.kt
@@ -2,7 +2,7 @@
  * Copyright 2018-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the MIT license.
  */
 
-@file:Suppress("UnstableApiUsage")
+@file:Suppress("UnstableApiUsage", "DialogTitleCapitalization")
 
 package com.microsoft.azure.toolkit.intellij.legacy.function.templates
 
@@ -15,7 +15,7 @@ import com.intellij.ui.dsl.builder.panel
 import com.intellij.util.ui.JBInsets
 import com.jetbrains.rider.ui.components.base.Viewable
 import com.microsoft.azure.toolkit.intellij.legacy.function.FUNCTIONS_CORE_TOOLS_LATEST_SUPPORTED_VERSION
-import com.microsoft.azure.toolkit.intellij.legacy.function.coreTools.FunctionCoreToolsInfoProvider
+import com.microsoft.azure.toolkit.intellij.legacy.function.coreTools.FunctionCoreToolsManager
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.swing.JComponent
@@ -39,11 +39,9 @@ class InstallFunctionToolComponent(reloadTemplates: Runnable) : Viewable<JCompon
                     val project = ProjectManager.getInstance().defaultProject
                     runWithModalProgressBlocking(project, "Downloading Azure Functions Core Tools...") {
                         withContext(Dispatchers.Default) {
-                            FunctionCoreToolsInfoProvider.getInstance()
-                                .retrieveForVersion(
-                                    FUNCTIONS_CORE_TOOLS_LATEST_SUPPORTED_VERSION,
-                                    true
-                                )
+                            FunctionCoreToolsManager
+                                .getInstance()
+                                .downloadLatestFunctionCoreToolsForVersion(FUNCTIONS_CORE_TOOLS_LATEST_SUPPORTED_VERSION)
                         }
                     }
                     reloadTemplates.run()